desafio_bycoders/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.3
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djlint==1.0.3
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.7
pathspec==0.9.0
psycopg2==2.9.3
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
zipp==3.8.0
